+++ packages/unlocked/tests/index.test.ts
@@ -1,16 +1,18 @@
 import { join } from "node:path";
 import { test, expect } from "vitest";
 
-import { parsePnpm } from "../lib";
-import { toCommonLockfile } from "../lib/lockfiles/pnpm";
+import { parse as parsePnpm, toCommonLockfile } from "../lib/lockfiles/pnpm";
 
 test("resolve-dependencies", async () => {
-	const lockfile = await parsePnpm(join(__dirname, "fixtures", "generic"));
+	let projectDirectory = join(__dirname, "fixtures", "generic");
+
+	const lockfile = await parsePnpm(projectDirectory);
 	expect(lockfile).toMatchSnapshot();
 
-	expect(
-		await toCommonLockfile(lockfile, {
-			projectDirectory: join(__dirname, "fixtures", "generic"),
-		}),
-	).toMatchSnapshot();
+	const commonLockfile = await toCommonLockfile(lockfile, {
+		packageJsonName: "package.json",
+		projectDirectory,
+	});
+
+	expect(commonLockfile).toMatchSnapshot();
 });