+++ packages/unlocked/CHANGELOG.md
@@ -1,5 +1,11 @@
 # template
 
+## 0.1.1
+
+### Patch Changes
+
+- fix broken tests
+
 ## 0.1.0
 
 ### Minor Changes